jl_rng_split: change p to w; regenerate some random constants
Rename `p` (for PCG) to `w` because we use the result as a weight in the
dot product construction. We're going to change the dot product to some
other combiner function, but `w` is still a better name.

I just happened to notice that the random additive constants were all
even which felt kind of weird so I just regenerated all of them. Since
we're changing the task fork result anyway, no extra disruption.